requirements.txt
fastapi>=0.104.0
pydantic>=2.4.0
requests>=2.31.0
tqdm>=4.65.0
uvicorn[standard]>=0.24.0
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0